# openapi/openapi.yaml
openapi: 3.1.1
info:
  title: QAIDILife Commerce API
  version: 1.0.0
  description: Core commerce API for QAIDILife storefront and mobile shells.
servers:
  - url: http://localhost:4000
    description: Local API
  - url: http://localhost:4010
    description: Mock API
paths:
...
  /v1/products:
    get:
      summary: List products
      operationId: listProducts
      parameters:
        - in: query
          name: size
          schema:
            type: string
        - in: query
          name: color
          schema:
            type: string
        - in: query
          name: price_min
          schema:
            type: number
        - in: query
          name: price_max
          schema:
            type: number
        - in: query
          name: sort
          schema:
            type: string
            enum: [newest, price_asc, price_desc, popular]
        - in: query
          name: cursor
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Product list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductList'
  /v1/products/{id}:
    get:
      summary: Get product by id or slug
      operationId: getProduct
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Product
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Product'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Money:
      type: object
      properties:
        currency:
          type: string
          default: BDT
        amount:
          type: number
      required: [currency, amount]
    ProductVariant:
      type: object
      properties:
        id:
          type: string
        sku:
          type: string
        size:
          type: string
        color:
          type: string
        stock:
          type: integer
        price:
          $ref: '#/components/schemas/Money'
      required: [id, sku, size, color, stock, price]
    Product:
      type: object
      properties:
        id:
          type: string
        slug:
          type: string
        title_en:
          type: string
        title_bn:
          type: string
        description_en:
          type: string
        description_bn:
          type: string
        primary_image:
          type: string
        gallery:
          type: array
          items:
            type: string
        price:
          $ref: '#/components/schemas/Money'
        compare_at:
          $ref: '#/components/schemas/Money'
        variants:
          type: array
          items:
            $ref: '#/components/schemas/ProductVariant'
        category_ids:
          type: array
          items:
            type: string
      required:
        - id
        - slug
        - title_en
        - title_bn
        - primary_image
        - price
        - variants
    ProductList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Product'
        next_cursor:
          type: string
          nullable: true
      required: [items]
...